Moved java version checker to standalone class 'JsonVersionLocator'; Added 'Java Executable' directory selector (doesn't save to file)

+++ APKognito/Cli/CliMain.cs
@@ -41,7 +41,7 @@ internal static class CliMain
             return;
         }
 
-        Exit((int)ExitCode.NoError);
+        Exit(ExitCode.NoError);
     }
 
     public static void AttachConsole()
